cmd/compile: improve debug locations partially live in-params
DO NOT REVIEW

During DWARF debug location generation, as a preamble to the main data
flow analysis, examine the function entry block to look for in-params
arriving in registers that are partially or completely dead, and
insert new OpArg{Int,Float}Reg values for the dead or part-dead
pieces. This helps create better/saner DWARF location expressions for
params. Example:

  func foo(s string, used int, notused int) int {
      return len(s) + used
  }

When optimization is complete for this function, the parameter
"notused" is completely dead, meaning that there is no entry for it in
the f.NamedValues table (which then means we don't emit a DWARF
variable location expression for it in the function enty block). In
addition, since only the length field of "s" is used, there is no
DWARF location expression for the other component of "s", leading to
pdegraded DWARF.

This are still problems/issues with DWARF location generation, but
this does improve the situation with respect to being able to print
the values of incoming parameters when stopped in the debugger at the
entry point of a function, for optimized code.

Change-Id: I5bb5253648942f9fd33b081fe1a5a36208e75785