[dev.fuzz] cmd/go/internal/cfg: enable "gofuzzbeta" tag by default
This lets users check in fuzz targets for use with this branch without
breaking the build for developers using a regular version of Go.

Before we merge this branch to master, this CL should be reverted. At
that point, users should change the tag to go1.18 (or whichever
version we land on).

Change-Id: I0e21a21e415e4fb7c599abe11e61de754c74a3d7