cmd/go: improve matching a file path in script tests
Currently, for portability reason, a path separator in a pattern must be
specified as '[.\\]', and a lot of tests need to match a file path.

Add the -path flag to the stdout, stderr and grep commands.  When it is
set, the UNIX path separator '/' is replaced with the OS specific path
separator before the match.

Change-Id: If535df95587332fe23b088dfc9e918d52674eb93