cmd/go: replace Target with MainModules, allowing for multiple targets
This change replaces the Target variable that represents the main module and
the pathPrefix and inGorootSrc which provide other information about the main
module with a single MainModules value that represents multiple main modules
and holds their path prefixes, module roots, and whether they are in
GOROOT/src. In cases where the code checks Target or its previously associated
variables, the code now checks or iterates over MainModules. In some cases, the
code still assumes a single main module by calling
MainModules.MustGetSingleMainModule. Some of those cases are correct: for
instance, there is always only one main module for mod=vendor. Other cases are
accompanied with TODOs and will have to be fixed in future CLs to properly
support multiple main modules.

This CL (and other cls on top of it) are planned to be checked into a branch to
allow for those evaluating the workspaces proposal to try it hands on.

For #45713

Change-Id: I3b699e1d5cad8c76d62dc567b8460de8c73a87ea