cmd/internal,cmd/link: add external linking support for SDYNIMPORT calls on mips64
Add external linking support for calling SDYNIMPORT symbols on
mips64 - when used with external linking, it is expected that the
caller will load register 25 (t9) with the address of the external
symbol using an offset from the global pointer (gp) register.

In order to achieve this, generate trampolines that have the correct
calling convention and appropriate relocations. It is expected that
the caller will have restored the gp register to the value initially
set by the dynamic linker and then restore it to Go's value following
the call.

Updates #36435
Fixes #46178

Change-Id: Idcd414a9a0e71604150b46b3a78fff56042fcbbb