[dev.fuzz] internal/fuzz: move coverage capture closer to function
When instrumented packages intersect with the packages used by the
testing or internal/fuzz packages the coverage counters become noisier,
as counters will be triggered by non-fuzzed harness code.

Ideally counters would be deterministic, as there are many advanced
fuzzing strategies that require mutating the input while maintaining
static coverage.

The simplest way to mitigate this noise is to capture the coverage
counters as closely as possible to the invocation of the fuzz target
in the testing package. In order to do this we need to export the
existing internal/fuzz.resetCoverage function, and add a new function
which captures the current values of the counters,
internal/fuzz.SnapshotCoverage. The latter function copies the current
counters into a static buffer, coverageSnapshot, which
workerServer.fuzz can then inspect when it comes time to check if new
coverage has been found.

This method is not foolproof. As the fuzz target is called in a
goroutine, harness code can cause counters to be incremented while the
target is being executed. Despite this we do see significant reduction
in churn via this approach. For example, running a  basic target that
causes strconv to be instrumented for 500,000 iterations causes ~800
unique sets of coverage counters, whereas by capturing the counters
closer to the target we get ~40 unique sets.

It may be possible to make counters completely deterministic, but
likely this would require rewriting testing/F.Fuzz to not use tRunner
in a goroutine, and instead use it in a blocking manner (which I
couldn't figure out an obvious way to do), or by doing something even
more complex.

Also, this uses a terrible linkname hack to avoid the complex import
cycle, probably we shouldn't do this. Perhaps a better solution would
be instead to move all of the coverage code into testing, with an
unexported API, and then linkname from internal/fuzz -> the unexported
testing stuff? Not really sure...

Change-Id: I95c2f3b1d7089c3e6885fc7628a0d3a8ac1a99cf