cmd/dist: exit early in bgwait after a fatal error
Currently, when a command scheduled by bgrun fails, bgwait returns to
the caller even in case the fatal error was detected.

In case of a syntax error in one of the standard packages, as an
example, the runInstall function will try to read the generated archive
file, only to fail since the file does not exist.  Since the runInstall
function is called in a goroutine, cmd/dist will continue to report
errors until all background goroutines are done.

Update the xexit function to call the atexit functions only once.  This
is necessary since xexit may be called multiple times in case of a fatal
error.  The bgwait function must exit the program, since returning to
the caller (with an error, as an example) will cause cmd/dist to report
an additional error during the next build phase.

Fixes #45410

Change-Id: If89976abad70f8d6ec79b2a5a1f2306e9c034c5a