runtime: fix types in openbsd/amd64 and openbsd/arm64 libc assembly
While working on the openbsd/mips64 libc conversion, it was discovered that
several types were incorrect in the openbsd/amd64 and openbsd/arm64 assembly
(which we get away with due to both being little endian). The type needs to
match that of the calling Go function, not the type of the libc syscall
(and currently they're not all the same types on both sides).

Updates #36435

Change-Id: I4b3dec8e88a33c8a35d9cc7a40562281b18c6487