run.bash: never pass the -no-rebuild flag to dist test
Currently, run.bash needs to pass the -rebuild flag to dist test and
all.bash needs to pass the -no-rebuild flag to run.bash in order to
prevent rebuilding.  In the end, the -no-rebuild flag needs to be
supported by dist test only for this use case.

Update the run.bash script to ensure that the -no-rebuild flag is never
passed to dist test, and instead is handled internally by run.bash to
unset the -rebuild flag.  Multiple -rebuild and -no-rebuild flags are
handled correctly.

Change-Id: I6fa29c30bf11381ce7e9f7b1a67681c2d8c77a0a