cmd/internal/obj/ppc64: add ISA 3.1 instructions
Cleanup ppc64 specific mapping of asm argument lists to obj.Prog
structures. In general, the final argument is p.From, and p.Reg
is skipped if the second arg is not a register type. Knowing this,
4-arg handling can be simplified and some helper functions removed.
Incorrect usage will also produce an error during argument matching
in asm9.go, this error will likely be more helpful.

Use ppc64map (from x/arch) to generate ISA 3.1 support for the
assembler. A new file asm9_gtables.go is added which contains
generated code to encode ISA 3.1 instructions, a function to assist
filling out the oprange structure, and a lookup table for the fixed
bits of each instructions. Generated functions are shared if their
bitwise encoding match, and the translation from an obj.Prog
structure matches.

Future improvements could shrink the generated opcode table
to 32 bit entries as there is much less variation of the
encoding of the prefix word, but it is not always identical
for instructions which share a similar encoding of arguments
(e.g PLWA and PLWZ).

Updates #44549

Change-Id: Ie83fa02497c9ad2280678d68391043d3aae63175