[dev.fuzz] internal/fuzz: notify coordinator for minimization
When a worker process finds a crasher, it now sends that result
directly to the coordinator without attempting to minimize it
first. The coordinator stops sending new inputs and sends the
unminimized crasher back to a worker (any worker) for minimization.

This prevents wasted work during minimization and will help us
implement -keepfuzzing later on. We may also be able to minimize
interesting inputs with this approach later.

DO NOT SUBMIT: wait for CL 321230 and rewrite test_fuzz_mutator.txt.
There will be fewer unrecoverable failures (since panics are
recoverable after that CL). We probably shouldn't attempt minimization
on unrecoverable failures, since the worker loses state when it
terminates.

Change-Id: Id142c7e91a33f64584170b0d42d22cb1f22a92d7