cmd/dist: don't use pathf when simply joining two paths
Add the join function to join two path segments, and use it instead of
pathf, when possible.

The filepath.Join function was not used, since in go1.4 it has some
potential problems on Windows.  In addition pathf was used 72% of the
time just to join two path segments, with only 3 times to join 3 path
segments.

Change-Id: I9486856a09e1634d1e1661726fb980ca65b36f7e