cmd/link: exit if cgo is required for external linking but is disabled
When cross compiling for a target requiring external linking, like
android and ios, the linker prints a message reporting that runtime/cgo
was not found and try to call the external linker, causing several
confusing errors to be reported to the user.

Update the Link.loadlib method to call Exitf in case cgo is required for
external linking but is not enabled.

Fixes #46330

Change-Id: Id7abc87c75809eb889fd11419c9d446f0fb4f53c