cmd/compile: do not mark stack object for map key passed to runtime function
When taking map key address to pass it to runtime function, we know that
the key is going to stay alive until the runtime function returns. So
even the key marked address taken, we don't need full generality of
stack object for it.

Instead, just emit OVARLIVE right after runtime function call, so the
liveness analysis and stack maps will take care of keeping the key
alive.

Fixed #43753

Change-Id: Ic2761c85f2ab88c0d81375eab5e577c1f03af2c0