[dev.typeparams] cmd/compile: lazy type definitions for types2
This CL extends the previous commit to also support lazy loading of
the underlying type and method set of defined types.

Nicely, this eliminates the awkwardness of needing to handle recursive
types within the importer. Instead, types2 is able to handle those
details.

The current prototype is a little fragile though. It works, but while
debugging, I kept running into issues with reentrant calls into
sync.Once.Do, which were a pain to debug. If we decide to do this for
real, we probably want to replace sync.Once with something that can
print useful diagnostics and stack traces rather than just
mysteriously hang.

Finally, it looks like the details are actually somewhat similar to
the lazy "instance" stuff used for generics, so maybe they can
actually use a common idea instead of duplicating the idea.

Updates #46449.

Change-Id: I7752c668a9169206318d4a777e4a0db2eb1e8ff7