cmd/compile: set "checkptr" to 2 when ASan is enabled
Set "-d=checkptr=2" when ASan is enabled, treats conversions
to unsafe.Pointer as an escaping operation. In this way, all
pointer operations on the stack objects will become operations
on the escaped heap objects. As we've already supported ASan
detection of error memory accesses to heap objects. With this
trick, we can use -asan option to report errors on bad stack
operations.

Add test cases.

Change-Id: I4e7fe46a3ce01f0d219e6a67dc50f4aff7d2ad87